rmでエラーになる

+++ scripts/lib/pyenv.sh
@@ -9,7 +9,9 @@ libbz2-dev libreadline-dev libsqlite3-dev curl llvm \
 libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev
 
 if [[ "$is_init" = true ]]; then
-    rm ~/.pyenv -rf
+    if [[ -e "$HOME/.pyenv" ]]; then
+        rm ~/.pyenv -rf
+    fi
 fi
 
 if [[ ! -e "$HOME/.pyenv" ]]; then
